Add `ColliderBackendPlugin`
