+++ crates/bevy_xpbd_3d/examples/custom_broad_phase.rs
@@ -13,7 +13,7 @@ fn main() {
     app.add_plugins(
         PhysicsPlugins::default()
             .build()
-            .disable::<BroadPhasePlugin<Collider>>()
+            .disable::<BroadPhasePlugin>()
             .add(BruteForceBroadPhasePlugin),
     );
